refactor: separate observer libs
